requirements.txt
matplotlib==3.6.3
openai==0.26.5
Pillow==9.4.0
streamlit==1.17.0
